Removed check on glyph true height
When extracting glyphs, top and bottom bearings are computed and
summed up to obtain the glyph height. This height may actually be lower
than the one specified for rendering (consider "short" characters like
'e' or 'm') or higher (consider "tall" characters like 'Ä' or 'Å'
which go past their bounding boxes). If the real height was lower than
specified the parser would crash lamenting an unexpected
error, so only blocks containing "big" characters would have made the
parsing successful, which is not a desired behaviour.