Restored Tahoma, fixed BDFParser issue
Tahoma bdf font doesn't have the standard replacement character (U+FFFD), but the
fallback one ('?') is also in the standard ASCII range. This means that, if not correctly
located and duplicated during the file scan of BDFParser, we have a missing glyph at the end
of the parsing.
Given this change, also the warning message of BDFParser has to be enriched.